+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable user-facing changes to this app will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+* Make it possible to view the previous photo on tap
+
 ## [v1.9.1] - 2024-06-11
 
 ### Fixed